refactor: move schedule list item dialog providers to util file
